+++ jolt-core/src/zkvm/instruction/virtual_sign_extend.rs
@@ -28,7 +28,7 @@ impl InstructionFlags for VirtualSignExtend {
 impl<const WORD_SIZE: usize> LookupQuery<WORD_SIZE> for RISCVCycle<VirtualSignExtend> {
     fn to_lookup_operands(&self) -> (u64, u128) {
         let (x, y) = LookupQuery::<WORD_SIZE>::to_instruction_inputs(self);
-        (0, x as u128 + y as u128)
+        (0, x as u128 + y as u64 as u128)
     }
 
     fn to_instruction_inputs(&self) -> (u64, i64) {
